Very basic solving func
Script takes string representing sudoku puzzle and attempts a solve.

Only currently does trivial checking of what numbers can fit in a box and if that list of numbers is length 1, puts the expected number in the box. Missing higher level intuitive solving of "is this the only square in box/row/col that can have number x?"

works for easy puzzles, not ones that are very much harder at all

very not great code. Could be saving state to minimize recalculations, and more hardcoded numbers than I'd like, but its a python script for fun with an n < 100 so I'm not too worried about it